% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCATE.R
\name{spCATE}
\alias{spCATE}
\title{Semiparametric targeted conditional average treatment effect estimation.
CATE(W) := E[Y|A=1,W] - E[Y|A=0,W]}
\usage{
spCATE(
  formula_CATE = ~1,
  W,
  A,
  Y,
  family_CATE = gaussian(),
  sl3_Lrnr_A = NULL,
  sl3_Lrnr_Y = NULL,
  weights = NULL,
  smoothness_order_Y0W = 1,
  max_degree_Y0W = 2,
  num_knots_Y0W = c(15, 5),
  fit_control = list()
)
}
\arguments{
\item{formula_CATE}{R-formula object specifying model for CATE}

\item{W}{A matrix of baseline covariates to condition on.}

\item{A}{A binary treatment assignment vector}

\item{Y}{An outcome variable (continuous or binary)}

\item{family_CATE}{A R-family object specifying the link function for the CATE}

\item{sl3_Lrnr_A}{An optional sl3-Learner object to estimate P(A=1|W)}

\item{sl3_Lrnr_Y}{An optional sl3-Learner object to estimate nuisance conditional means E[Y|A=0,W] and E[Y|A=1,W]}

\item{weights}{A vector of optional weights.}

\item{smoothness_order_Y0W}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}

\item{max_degree_Y0W}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}

\item{num_knots_Y0W}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}

\item{fit_control}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}
}
\description{
Semiparametric targeted conditional average treatment effect estimation.
CATE(W) := E[Y|A=1,W] - E[Y|A=0,W]
}
