% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spRR.R
\name{spRR}
\alias{spRR}
\title{Semiparametric targeted conditional relative risk/treatment effect using machine-learning
RR(W) := E[Y|A=1/W] / E[Y|A=0,W]}
\usage{
spRR(
  formula_logRR = ~1,
  W,
  A,
  Y,
  family_RR = gaussian(),
  sl3_Lrnr_A = NULL,
  sl3_Lrnr_Y = NULL,
  weights = NULL,
  smoothness_order = 1,
  max_degree = 2,
  num_knots = c(15, 5),
  fit_control = list()
)
}
\arguments{
\item{formula_logRR}{R-formula object specifying model for log relative risk}

\item{W}{A matrix of baseline covariates to condition on.}

\item{A}{A binary treatment assignment vector}

\item{Y}{A nonnegative outcome variable. Can be binary, a count, or a continuous nonnegative variable.}

\item{family_RR}{A R-family object specifying the link function for the log relative risk (gaussian/identity implies formula_logRR is directly modelling the log RR)}

\item{sl3_Lrnr_A}{An optional sl3-Learner object to estimate P(A=1|W)}

\item{sl3_Lrnr_Y}{An optional sl3-Learner object to estimate nuisance conditional means E[Y|A=0,W] and E[Y|A=1,W]}

\item{weights}{A vector of optional weights.}

\item{smoothness_order}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}

\item{max_degree}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}

\item{num_knots}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}

\item{fit_control}{Specification for default HAL learner (used if sl3 Learners not given). See spOR for use.}
}
\description{
Semiparametric targeted conditional relative risk/treatment effect using machine-learning
RR(W) := E[Y|A=1/W] / E[Y|A=0,W]
}
